remove Input from ol-components
